requirements.txt
fastapi==0.100.1
numpy==1.26.0
python-dotenv==1.0.1
requests==2.28.2
uvicorn[standard]==0.23.2